Servidor/requirements.txt
openai==1.82.1
python-dotenv==1.1.0
flask==3.1.1
pymysql==1.1.1